requirements.txt
asgiref==3.4.1
authy==2.2.6
autopep8==1.6.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.10
Django==3.2
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.4
idna==3.3
mysqlclient==2.1.0
phonenumbers==8.12.21
Pillow==9.0.0
pycodestyle==2.8.0
PyJWT==1.7.1
PyMySQL==1.0.2
python-decouple==3.5
python-dotenv==0.17.0
pytz==2021.3
razorpay==1.2.0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
twilio==6.57.0
typing-extensions==4.0.1
urllib3==1.26.8
xlwt==1.3.0
